+++ lesson8/src/granny-smith.ts
@@ -1,0 +1,11 @@
+import { Apples } from './abstraction';
+
+export class GrannySmith extends Apples {
+    public constructor() {
+        super("tart", "green");
+    }
+
+    public getAppleType(): string {
+        return "Granny Smith";
+    }
+}
