+++ lesson8/src/red-chief.ts
@@ -1,0 +1,11 @@
+import { Apples } from './abstraction';
+
+export class RedChief extends Apples {
+    public constructor() {
+        super("sweet", "red");
+    }
+
+    public getAppleType(): string {
+        return "Red Chief";
+    }
+}
